Moved GEMM kernel predicates to separate files
